update type value in data source and classes

+++ src/main/java/com/CezaryZal/entity/score/Dimension.java
@@ -2,11 +2,11 @@ package com.CezaryZal.entity.score;
 
 public class Dimension {
 
-    private double valueOfDimension;
+    private int valueOfDimension;
     private double lowerDeviation;
     private double upperDeviation;
 
-    public Dimension(double valueOfDimension, double lowerDeviation, double upperDeviation) {
+    public Dimension(int valueOfDimension, double lowerDeviation, double upperDeviation) {
         this.valueOfDimension = valueOfDimension;
         this.lowerDeviation = lowerDeviation;
         this.upperDeviation = upperDeviation;
